+++ app/src/main/java/com/example/at/models/TrainingModel.kt
@@ -1,13 +1,10 @@
 package com.example.at.models
 
-import android.provider.ContactsContract
 import com.jjoe64.graphview.series.DataPoint
-import java.time.Duration
-import java.time.LocalDateTime
 
 class TrainingModel {
-    var timeStart = 0
-    var duration = 0
+    var timeStart = 0.toLong()
+    var duration = 0.toLong()
 
     var x: Int = 0
     var y: Int = 0
